fix mysql bug with DateTime
PDO can't bind a DateTime parameter: it needs to be explicitely turned into
string. This bug hasn't been spotted earlier because this method is currently
only called by the endpoint MAWhoNeedToRenew.php which passes a string. The
caller is the class OutdatedMemberDeleter which uses a DateTime but this code
runs only once a year (when the old registrations are cleaned) and it hasn't
run yet.