Fix mysql bug
the commit which introduced the debug flag also introduced a regression
because the parameters weren't passed anymore to the statement which
updates the status related to "notification has been sent to admins".

Nb: It's the only statement where parameters are passed through the
`execute` method (rather than using `bindParam`). (That's because the
number of parameter is variable). That's hence the only place to fix