Don't fail on donations
It is now possible to make donations with the current membership forms.
When a user register and make a donation, the API call return 2 different lines
of data, one for each. Since the point of those scripts is to track only
registrations we filter those out (without this patch the runs fail because
we're not able to parse data for donations (which makes sense since we don't
want to parse it))