Change the license for GPL v3
This will make it more convenient to be compatible with the submodule
slack-app which is going to become GPLP v3 soon.

For simplicity I also change the license of the directory files/google;
since it's was under "Do What The F*uck You Want To Public License", my
understanding is that it is possible to relicense this just like that.

Nb: we're going to merge this commit through a github pull request in
order to gather the approval of people who contributed to this repo