Disable usage of http-mock
that package is not compatible with PHP8. We opened an issue
related to that almost a year ago (see
https://github.com/InterNations/http-mock/issues/69 ), and we tried
updating it ourselves, but it appears to be non trivial.

Since I'm now not able to run those tests on my laptop (since I
upgraded to ubuntu 22.04) I'm deactivating the associated tests
(not completely deleting them for now as I still hope it will
be possible to reactivate them soon and since I'm afraid deleting
them would just lead to completely forget them)