Don't send technical logs to the browser
Now the db connector logs stuff when membersPerPostalCode is displayed.
I introduce this Noop logger as a quick fix.

(Ultimately a better approach could be to be able to log only in files,
and to ensure we use different log files for the web page and for the
cron script.
However it would also make more sense to use an actual 3rd party logging
lib. We'll take the time to properly implement all this once the current
work is done)