Fix bug in case where HA events aren't received in asc. order
It ocurred for instance when I did a bug since the beginning of
the year in order to fill the table handled by Doctrine.
What happened is that we handled the more recent registration first
and the older afterwards, which ended up overriding the more recent data.

This commit fixes this behavior.