Update README
Document how to create the new member table instead of the old.
Also remove a part of the README which was not very useful and which
could easily get out of date.
And remove the "TODO" part which is not so useful anymore given
the plans we have for this repo anyway.