Add GroupMemberDeleter
After some tests on actual data it appears that:
- the case may not be the same for some email adress between
  what they are on helloasso and what they are when retrieved
  afterward from either the google group or mailchimp.
  It's not clear why, but it's what we observe
- we can delete an email from the google group without matching
  the exact case, but it does not work for mailchimp (which
  requires the exact case known by it)

so our code needed to be a little more complex, in order to match
user with case insensitive comparison in order to find out which
emails to delete; but then to perform the deletion for the exact
case known by the group.

This commit:
- fixes this deletion code
- extracts it in an ad hoc class (because MemberImporter is already
  pretty complex, and because it makes that logic easier to test)

It also changes GroupWithDeletableUsers interface so it can delete
single users instead of list of users. That's because in practice
we don't have any group that can do a batch deletion, so it's pointless
to aggregate users in an array before looping on it afterwards.