add test_deleteOutdatedMembers and fix this code
writting this test revealed a small bug: we kept in the groups
all the members still in db, though the db is cleaned after a year,
which means we kept users in groups for too long.
(not that it's not an issue in prod because this code was deployed
 after the last batch of clean-up).

This commmit also changes a bit the GroupMemberDeleter interface
in order to make it more convenient to test.