Add openapi-login.yaml
It would be more convenient to be able to use generated
code to perform authentication queries from the front.
But OTHO we don't want to generate symfony code for those
endpoints since authentication is handled by Symfony itself.
So I add those endpoints in a separated openapi file.

Also use a more meaningful field in the response